backend/requirements.txt
langgraph
langgraph-checkpoint-sqlite
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
pytest
pytest-asyncio
openai
python-dotenv
